fix![dependencies] Fixed various errors detected by intellij and added dependencies recommended from code review

+++ src/main/java/com/example/user/domain/repository/UserRepository.java
@@ -8,7 +8,7 @@ public interface UserRepository {
 
     User findOne(String email);
 
-    int insert(User user);
+    long insert(User user);
 
     void delete(String email);
