+++ src/main/java/com/example/user/domain/service/UserService.java
@@ -4,5 +4,5 @@ import com.example.user.domain.model.User;
 import com.example.user.impl.domain.UserAlreadyExistAuthenticationException;
 
 public interface UserService {
-    int registerNewUserAccount(User user) throws UserAlreadyExistAuthenticationException;
+    long registerNewUserAccount(User user) throws UserAlreadyExistAuthenticationException;
 }
